// src/app/models/master-dealer-store-form.model.ts
export interface MasterDealerStoreFormModel {
  id: number;
  companyName: string;
  cnpj: string;
  commercialname: string;
  postalCode: string;
  address: string;
  addressNumber: string;
  complement: string;
  neighborhood: string;
  city: string;
  state: string;
  phone: string;
  email: string;
  site: string;
  image: string;
}